requirements.txt
python-crontab~=3.3.0
sqlalchemy~=2.0.48
psycopg2-binary~=2.9.11
pandas~=2.3.3
requests~=2.32.5
beautifulsoup4~=4.14.3
google-auth~=2.49.1
google-api-python-client~=2.193.0
gspread~=6.1.4  
pytest~=9.0.2
pytest-cov~=6.3.0